Merge branch 'python3' of https://github.com/MrSenko/python-nitrate
this adds a copy of the original python-nitrate API client with
changes that make it possible to execute on Python 3.

Due to incompatibilities in config file names and config sections
we are forking the original API client and will be developing it
in this repository. This will also have positive effects on testing
both the API client and the XML-RPC API server because we can do it
more easily.

Also we're planning to remove some of the original pieces, most
notably Teiid support!