Update xmlrpc.py
To support using winkerberos instead of kerberos module when running the tcms-api on Microsoft Windows platform.

Background: 
The kerberos module does not support Microsoft Windows platform. We need to replace it with the winkerberos.
While there is a workaround in the pip installation, but the code itself does not detect the platform and use the correct module.
Adding the above would try to import kerberos first, and if failed, will use winkerberos.