Fix failing kerberos authentication. Fixes #22
Kiwi TCMS (via social_auth_kerberos) will only check tokens during
the login process and afterwards rely on the session cookie!

Because of this just simulate a call to /login/kerberos/ and if
successfull set the session cookie on all subsequent requests.

The original behavior of sending the authorization request header
is preserved in case someone is using Apache mod_kerb instead of
PSA Kerberos backend.