Start building packages on Windows. Fix #3
NOTE: the wheel package itself is a pure-python package so it
will not matter on what OS it has been built! With the latest
changes to use winkerberos over kerberos customers should be able
to pip install directly on any operating system and pip will
resolve the dependencies for them.

- Prepend C:\cygwin64\bin to PATH
- Prepend C:\Users\appveyor\AppData\Roaming\Python\Python36\Scripts
  to PATH (for wheel.exe)
- unpack wheel packages instead of using unzip -t b/c in AppVeyor
  cygwin64 doesn't have unzip and the one which is in PATH fails
- Use the default Python executable when building venvs for testing
- Replace "bin/activate" with "Scripts/activate" via sed
- Remove "pip install kerberos" b/c winkerberos ships both Source and
  Whell packages
- Remove "-p $(which python)" when calling virtualenv b/c it fails. In
  AppVeyor we have C:\Python36-x64\Scripts\virtualenv.exe in PATH
- Start keeping the built artifacts on AppVeyor