Use current user if plugin_helpers.Backend.default_tester_id is None
default_tester_id may still be specified via ENV or via inherited
classes but the code is now more tolerant. If the resulting value is
None then simply use the user making the API request (which may be a
different user, e.g. an automation integration account).

This commit removes the call to `User.filter` API method for which most
users may not be authorized!