Refactor calling .login() outside of __init__() method of Xmlrpc classes
and call .login() as part of TCMS.__init__() instead!

`__init__` method calls overridden method :

Call to self.login in __init__ method, which is overridden by method
TCMSKerbXmlrpc.login.

When an instance of a class is initialized, the super-class state should
be fully initialized before it becomes visible to the subclass. Calling
methods of the subclass in the superclass' __init__ method violates this
important invariant.